# go/go/src/ipsap/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "IPSAP Service APIs",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "basePath": "/api/v1.0",
    "paths": {
...
        "/orders": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "결제 이력",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "결제 이력",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "기관 seq",
                        "name": "filter.institution_seq",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {}
                }
            },
            "post": {
                "description": "상품 결제",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "상품 결제",
                "responses": {
                    "200": {}
                }
            }
        },
        "/orders/assign": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "상품 결제창 요청",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "상품 결제창 요청",
                "parameters": [
                    {
                        "description": "order assign",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.orderAssginModel"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/orders/{order_seq}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "결제 상세",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "결제 상세",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "1",
                        "name": "order_seq",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            },
            "delete": {
                "description": "결제 취소 (플랫폼 관리자) 기관에 상태 변경 적용 안됨\npartial_cancel_code 0:전체 취소, 1:부분 취소",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "결제 취소 (플랫폼 관리자)",
                "parameters": [
                    {
                        "description": "order cancel",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.orderCancelModel"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "1",
                        "name": "order_seq",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
...
    },
    "definitions": {
...
        "model.orderAssginModel": {
            "type": "object",
            "properties": {
                "amt": {
                    "type": "integer",
                    "example": 1000000
                },
                "edi_date": {
                    "type": "string",
                    "example": "YYYYMMDDHHMMSS"
                },
                "goods_name": {
                    "type": "string",
                    "example": "정기 1개월권"
                },
                "product_seq": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "model.orderCancelModel": {
            "type": "object",
            "properties": {
                "cancel_amt": {
                    "type": "string",
                    "example": "1000"
                },
                "partial_cancel_code": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}